Added new class and list
Moved the scroll-content styling to app.scss styling, so that it is
everywhere on the app.
Added some pictures to the lists.
Added the new class object to the guide class.
Looped through the array in guide-lists.html.
Made a method that send the user to the new list, with the selectedList